Add --format option
Adds a --format (-f) option to control the output format of errors and
warnings. Valid values for --format are 'text' (the default) and 'json'.
The 'text' format is similar to the previous format but slightly more
structured. 'json' format is brand new.

Tests have been updated to match against the slight changes to 'text'
output format as well as resolving some minor inconsistencies in which
checks would result in a non-zero exit status.

Due to the more consistent behavior around exit status, these changes
are not strictly backwards compatible.